auth-service/pyproject.toml
[project]
name = "codelab-auth-service"
version = "0.1.0"
description = "CodeLab Auth Service - OAuth2 Authorization Server"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Sergey Penkovsky", email = "sergey.penkovsky@gmail.com" }]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.1",
    "pydantic==2.5.1",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "aiosqlite==0.19.0",
    "redis==5.0.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "python-multipart==0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio", "pytest-cov"]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
